requirements.txt
absl-py==1.3.0
aiohttp==3.8.3
aiosignal==1.3.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.2.0
cachetools==5.2.0
catboost==1.1.1
certifi==2022.12.7
charset-normalizer==2.1.1
cloudpickle==2.2.0
cmdstanpy==1.0.8
colorama==0.4.6
contourpy==1.0.6
convertdate==2.4.0
cycler==0.11.0
Cython==0.29.32
docutils==0.19
ephem==4.1.4
flatbuffers==22.12.6
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2022.11.0
gast==0.4.0
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.20.1
grpcio==1.51.1
h5py==3.7.0
hijri-converter==2.2.4
holidays==0.18
idna==3.4
joblib==1.2.0
keras==2.11.0
kiwisolver==1.4.4
korean-lunar-calendar==0.3.1
libclang==14.0.6
lightgbm==3.3.4
lightning-utilities==0.5.0
llvmlite==0.39.1
LunarCalendar==0.0.9
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
multidict==6.0.4
nfoursid==1.0.1
numba==0.56.4
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==22.0
pandas==1.5.2
patsy==0.5.3
Pillow==9.3.0
plotly==5.11.0
pmdarima==2.0.2
prophet==1.1.1
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyMeeus==0.5.12
pyod==1.0.7
pyparsing==3.0.9
python-dateutil==2.8.2
pytorch-lightning==1.8.6
pytz==2022.7
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.0
scipy==1.9.3
setuptools-git==1.2
shap==0.41.0
six==1.16.0
slicer==0.0.7
statistics==1.0.3.5
statsforecast==1.4.0
statsmodels==0.13.5
tbats==1.1.2
tenacity==8.1.0
tensorboard
tensorboard-data-server
tensorboard-plugin-wit
tensorboardX
tensorflow
tensorflow-estimator
tensorflow-intel
tensorflow-io-gcs-filesystem
termcolor==2.1.1
threadpoolctl==3.1.0
torch==1.13.1
torchmetrics==0.11.0
tqdm==4.64.1
typing_extensions==4.4.0
u8darts==0.23.0
urllib3==1.26.13
Werkzeug==2.2.2
wrapt==1.14.1
xarray==2022.12.0
xgboost
yarl==1.8.2
